moves mga functions to utils
