+++ osier/__init__.py
@@ -15,6 +15,5 @@ from .models.capacity_expansion import *
 from .models.deap_runner import *
 from .utils import *
 from .equations import *
-from .nmga import nmga
 
 from .tech_library import *
